pyproject.toml
[project]
name = "io_scene_mw"
version = "0.8.112"
description = "Morrowind Blender Plugin"
authors = [{ name = "Greatness7" }]
dependencies = []
requires-python = ">=3.11, <3.12"
readme = "README.md"

[dependency-groups]
dev = [
    "numpy (>=1.24, <1.25)",
    "fake-bpy-module-latest",
]
